requirements.txt
web3
python-dotenv
asyncio
asyncio-queue